sc2 (#32): Launching the game via sc2 switcher rather than sc2 bot

+++ worlds/sc2/apclient/banks.py
@@ -4,7 +4,8 @@ import logging
 import os
 import queue
 import re
-import sys
+from . import user_paths
+from .failable import Error
 
 # Banks are XML files used to communicate with Starcraft 2
 # file names, section names and key names have to match the SC2 trigger implementation
@@ -65,60 +66,6 @@ BANK_BACKUP_FILE_LIMIT = 100
 
 logger = logging.getLogger("Starcraft2")
 
-def _get_bank_folder() -> str:
-    DOCUMENTS_SC2_DIRNAME = "Documents/StarCraft II"
-    BANKS_DIRNAME = "Banks"
-
-    # Environment variable override
-    user_set = os.environ.get("SC2_DOCUMENTS_DIR")
-    if user_set is not None:
-        if not os.path.isdir(user_set):
-            raise ValueError(f"User-defined folder {user_set} doesn't exist or is not a folder")
-        result = os.path.join(user_set, BANKS_DIRNAME)
-        if os.path.isfile(result):
-            raise ValueError(f"{result} is a file; expected a folder")
-        if not os.path.isdir(result):
-            os.makedirs(result)
-        return user_set
-
-    # Windows handling
-    if sys.platform == 'win32':
-        # Handle documents folder backed up by cloud service (OneDrive)
-        banks_folders = glob(os.path.expanduser(f"~/*/{DOCUMENTS_SC2_DIRNAME}/{BANKS_DIRNAME}"))
-        if len(banks_folders) > 0:
-            result = banks_folders[0]
-        else:
-            result = os.path.expanduser(f"~/{DOCUMENTS_SC2_DIRNAME}/{BANKS_DIRNAME}")
-        assert os.path.isdir(result), "Banks folder not found"
-        return result
-    
-    # Linux handling
-    wine_prefix = os.environ.get("WINEPREFIX")
-    if wine_prefix is not None:
-        user_folder_name = os.path.basename(os.path.expanduser("~"))
-        sc2_documents_dir = os.path.join(wine_prefix, f"drive_c/users/{user_folder_name}/{DOCUMENTS_SC2_DIRNAME}")
-    else:
-        sc2_documents_dir = os.path.expanduser(f"~/{DOCUMENTS_SC2_DIRNAME}")
-    result = os.path.join(sc2_documents_dir, BANKS_DIRNAME)
-    if not os.path.isdir(sc2_documents_dir):
-        raise ValueError(
-            f"sc2 documents folder {sc2_documents_dir} does not exist. "
-            f"Ensure WINEPREFIX or SC2_DOCUMENTS_DIR are set correctly"
-        )
-    if os.path.isfile(result):
-        raise ValueError(f"{result} is a file; expected a folder")
-    if not os.path.exists(result):
-        os.makedirs(result)
-    return result
-
-
-_sc2_bank_folder: str | None = None
-def get_bank_folder() -> str:
-    global _sc2_bank_folder
-    if _sc2_bank_folder is None:
-        _sc2_bank_folder = _get_bank_folder()
-    return _sc2_bank_folder
-
 
 class SC2Bank:
     # Has the same structure as bank files provided by SC2
@@ -151,16 +98,19 @@ class SC2Bank:
                 result = self.sections[section][key]
         return result
 
-    def read_file(self, path: str = "") -> None:
+    def read_file(self, path: str = "") -> None | Error[str]:
         """
         Reads a bank file provided by SC2 and converts it into an SC2Bank object.
         Assumes bank files follow the structure provided by the game.
         Asserts catch malformed files, which should never happen unless the player manually edits the files.
         """
         if not path:
-            path = f"{get_bank_folder()}/{self.file_name}.SC2Bank"
+            bank_folder = user_paths.get_bank_folder()
+            if isinstance(bank_folder, Error):
+                return bank_folder
+            path = f"{bank_folder}/{self.file_name}.SC2Bank"
         if not os.path.isfile(path):
-            return
+            return None
         with open(path, "r") as f:
             SECTION_PATTERN = re.compile(r'<Section name="(\w+)">')
             KEY_PATTERN = re.compile(r'<Key name="(\w+)">')
@@ -175,13 +125,13 @@ class SC2Bank:
                 if (m := SECTION_PATTERN.match(line_content)):
                     assert not section, f"Encountered section {m.group(1)} while already inside section {section}"
                     section = m.group(1)
+                    assert section not in self.sections, f"Duplicate section definition for section {section}"
                     self.add_section(section)
-                    assert section not in self.sections.items(), f"Duplicate section definition for section {section}"
                 elif (m := KEY_PATTERN.match(line_content)):
                     assert section, "Encountered a key while not in a section"
                     assert not key, f"Encountered key {m.group(1)} while already inside key {key}"
                     key = m.group(1)
-                    assert key not in self.sections[section].items(), f"Duplicate key definition for key {key}"
+                    assert key not in self.sections[section], f"Duplicate key definition for key {key}"
                 elif (m := VALUE_PATTERN.match(line_content)):
                     assert section, "Encountered a value while not in a section"
                     assert key, "Encountered a value while not in a key"
@@ -195,12 +145,16 @@ class SC2Bank:
                     section = ''
                 else:
                     pass
+        return None
 
-    def write_file(self) -> None:
+    def write_file(self) -> None | Error[str]:
         # Write a bank file with the formatting expected from SC2
         # SC2 reads data by restoring a bank from a backup
         # so we write the new backup file here
-        dir = f"{get_bank_folder()}/Backup"
+        bank_folder = user_paths.get_bank_folder()
+        if isinstance(bank_folder, Error):
+            return bank_folder
+        dir = f"{bank_folder}/Backup"
         lines: list[str] = []
         lines.append(         f'<?xml version="1.0" encoding="utf-8"?>')
         lines.append(         f'<Bank version="1">')
@@ -221,14 +175,19 @@ class SC2Bank:
             if not os.path.isfile(path):
                 with open(path, "w") as f:
                     f.write('\n'.join(lines))
-                return
+                return None
         # only the messages bank has any chance to hit this
         # at current limits, this would be attempting to send 5000 messages in a single iteration
         logger.info(f"Too many bank backups, cannot write:\n{self}")
+        return None
 
-    def remove_entry_from_file(self, key: str) -> None:
-        # Remove one Key/Value pair from a bank file
-        path = f"{get_bank_folder()}/{self.file_name}.SC2Bank"
+    def remove_entry_from_file(self, key: str) -> None | Error[str]:
+        """Remove one Key/Value pair from a bank file"""
+
+        bank_folder = user_paths.get_bank_folder()
+        if isinstance(bank_folder, Error):
+            return bank_folder
+        path = f"{bank_folder}/{self.file_name}.SC2Bank"
         with open(path, "r") as f:
             lines = f.readlines()
         with open(path, "w") as f:
@@ -243,30 +202,38 @@ class SC2Bank:
                     deleting = False
                 elif not deleting:
                     f.write(line)
+        return None
 
 
-def file_cleanup() -> None:
+def file_cleanup() -> None | Error[str]:
     # Use at the start of a mission to delete old bank files
     # Locations needs cleanup, others are optional
-    Path(f"{get_bank_folder()}/{BANK_LOCATIONS_FILE_NAME}.SC2Bank").unlink(missing_ok=True)
+    bank_folder = user_paths.get_bank_folder()
+    if isinstance(bank_folder, Error):
+        return bank_folder
+    Path(f"{user_paths.get_bank_folder()}/{BANK_CORE_OPTIONS_FILE_NAME}.SC2Bank").unlink(missing_ok=True)
+    Path(f"{user_paths.get_bank_folder()}/{BANK_LOCATIONS_FILE_NAME}.SC2Bank").unlink(missing_ok=True)
+    Path(f"{user_paths.get_bank_folder()}/{BANK_TRADE_RECEIVE_FILE_NAME}.SC2Bank").unlink(missing_ok=True)
+    Path(f"{user_paths.get_bank_folder()}/{BANK_MESSAGES_FILE_NAME}.SC2Bank").unlink(missing_ok=True)
+    return None
 
 
-def send_options(msg: str) -> None:
+def send_options(msg: str) -> None | Error[str]:
     bank = SC2Bank(BANK_OPTIONS_FILE_NAME)
     bank.add_entry(
         BANK_OPTIONS_SECTION_OPTIONS,
         BANK_OPTIONS_KEY_OPTIONS,
         msg
     )
-    bank.write_file()
+    return bank.write_file()
 
 
 def send_core_options(
     start_resources: str,
     colors: str,
-    uncollected_objectives: str = None,
-    finished_loading: str = None
-) -> None:
+    uncollected_objectives: str | None = None,
+    finished_loading: str | None = None
+) -> None | Error[str]:
     bank = SC2Bank(BANK_CORE_OPTIONS_FILE_NAME)
     bank.add_entry(
         BANK_CORE_OPTIONS_SECTION_CORE_OPTIONS,
@@ -290,7 +257,7 @@ def send_core_options(
             BANK_CORE_OPTIONS_KEY_LOAD_FINISHED,
             finished_loading
         )
-    bank.write_file()
+    return bank.write_file()
 
 
 def send_items(
@@ -298,7 +265,7 @@ def send_items(
     zerg_items: str,
     protoss_items: str,
     misc_items: str,
-) -> None:
+) -> None | Error[str]:
     bank = SC2Bank(BANK_ITEMS_FILE_NAME)
     bank.add_entry(
         BANK_ITEMS_SECTION_ITEMS,
@@ -320,14 +287,14 @@ def send_items(
         BANK_ITEMS_KEY_MISC_ITEMS,
         misc_items
     )
-    bank.write_file()
+    return bank.write_file()
 
 
 def escape_ap_message(message: str) -> str:
     return message.replace('&', '&amp;').replace('<', '&lt;').replace('>', '&gt;').replace('"', '&quot;')
 
 
-def send_ap_messages_from_queue(message_queue: queue.Queue):
+def send_ap_messages_from_queue(message_queue: queue.Queue) -> None | Error[str]:
     messages = []
     for i in range(BANK_MESSAGES_KEY_LIMIT):
         # send up to KEY_LIMIT messages in a single bank file
@@ -337,10 +304,11 @@ def send_ap_messages_from_queue(message_queue: queue.Queue):
         else:
             break
     if messages:
-        send_ap_message(messages)
+        return send_ap_message(messages)
+    return None
 
 
-def send_ap_message(messages: list[str]):
+def send_ap_message(messages: list[str]) -> None | Error[str]:
     # message list is expected to not contain more than KEY_LIMIT messages
     bank = SC2Bank(BANK_MESSAGES_FILE_NAME)
     if messages:
@@ -354,12 +322,17 @@ def send_ap_message(messages: list[str]):
                     escape_ap_message(msg)
                 )
         if i > 0:
-            bank.write_file()
+            return bank.write_file()
+    return None
 
 
 def update_prompt() -> bool:
+    """Checks to see if the game has requested an update by writing the ArchipelagoUpdate file"""
+    bank_folder = user_paths.get_bank_folder()
+    if isinstance(bank_folder, Error):
+        return False
     result = False
-    path = f"{get_bank_folder()}/{BANK_UPDATE_FILE_NAME}.SC2Bank"
+    path = f"{user_paths.get_bank_folder()}/{BANK_UPDATE_FILE_NAME}.SC2Bank"
     if os.path.isfile(path):
         # if bank exists, we want an update prompt. No need to check the values
         os.remove(path)
@@ -367,9 +340,10 @@ def update_prompt() -> bool:
     return result
 
 
-def read_locations() -> str:
+def read_locations() -> str | Error[str]:
     bank = SC2Bank(BANK_LOCATIONS_FILE_NAME)
-    bank.read_file()
+    if error := bank.read_file():
+        return error
     result = bank.get_value(
         BANK_LOCATIONS_SECTION_LOCATIONS,
         BANK_LOCATIONS_KEY_GAME_STATE
@@ -378,35 +352,41 @@ def read_locations() -> str:
 
 
 # Void Trade
-def send_trade_received(msg: str) -> None:
+def send_trade_received(msg: str) -> None | Error[str]:
     bank = SC2Bank(BANK_TRADE_RECEIVE_FILE_NAME)
     bank.add_entry(
         BANK_TRADE_RECEIVE_SECTION_TRADE,
         BANK_TRADE_RECEIVE_KEY_TRADE_RESPONSE,
         msg
     )
-    bank.write_file()
+    return bank.write_file()
 
 
-def read_trade_units() -> str:
+def read_trade_units() -> str | Error[str]:
     bank = SC2Bank(BANK_TRADE_SEND_FILE_NAME)
-    bank.read_file()
+    if error := bank.read_file():
+        return error
     result = bank.get_value(
         BANK_TRADE_SEND_SECTION_UNITS,
         BANK_TRADE_SEND_KEY_UNIT_TYPES
     )
     if result:
-        bank.remove_entry_from_file(BANK_TRADE_SEND_KEY_UNIT_TYPES)
+        error = bank.remove_entry_from_file(BANK_TRADE_SEND_KEY_UNIT_TYPES)
+        if isinstance(error, Error):
+            return error
     return result
 
 
-def read_trade_receive_request() -> str:
+def read_trade_receive_request() -> str | Error[str]:
     bank = SC2Bank(BANK_TRADE_SEND_FILE_NAME)
-    bank.read_file()
+    if error := bank.read_file():
+        return error
     result = bank.get_value(
         BANK_TRADE_SEND_SECTION_RECEIVE_REQUEST,
         BANK_TRADE_SEND_KEY_RECEIVE_COUNT
     )
     if result:
-        bank.remove_entry_from_file(BANK_TRADE_SEND_KEY_RECEIVE_COUNT)
+        error = bank.remove_entry_from_file(BANK_TRADE_SEND_KEY_RECEIVE_COUNT)
+        if isinstance(error, Error):
+            return error
     return result
